Rename `service_model` to `service`
