+++ lib/menu.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
-import 'package:ten_ten_one/model/service_model.dart';
+import 'package:ten_ten_one/model/service.dart';
 
 class Menu extends StatelessWidget {
   const Menu({Key? key}) : super(key: key);
